mini: add support for drive fallocate/truncate
Allow to set this in config and update runner to use it.

Fallocate makes sure the whole size could be used eventually, so there
are no spurious ENOSPC errors during the VM run. Lack of space is
detected at the machine run time if the 'cleandisk' option is set.

Truncate uses less space if the images are only sparsely used but may
fail if there's not enough space for all drives at some point.

Signed-off-by: David Sterba <dave@jikos.cz>