mini: fix installation patterns and substitutions
There are package rquirements that may be satisfied by various packages
depending on the patterns. When there's minimal system the busybox
substitutes take precedence and install e.g. gzip, sed, grep. But
they're not 1:1 compatible with the proper utilities.

Add support for forced uninstallation (prepend ! to the package name),
which will suggest the alternative packages.

Signed-off-by: David Sterba <dave@jikos.cz>