mini: use explicit loop device for root mount
Starting in some 6.x version the automatic loop device assigned by
'mount -o loop' was impossible to delete by 'losetup -d'. This caused
cortuptions after updating distro or kernel. 'fsck' was able to fix most
of that but that's not reliable.

Work around that by allocating the loop device before mount, but reuse
the same device if there's one.

Signed-off-by: David Sterba <dave@jikos.cz>